+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+---
+
+## \_\_-Sep-2025
+
+### Added:
+
+- App set to a multipage app with top navigation.
+- App 1: Conference Booking
+- App 2: Request Software
+- Added contact details in the email body.
+
+### Changed:
+
+- Improved the TIP section layout and its content.
+
+### Fixed
+
+- ***
+
 ## 17-Sep-2025
 
 ### Added:
@@ -11,7 +30,7 @@
 - Success flash position changed to bottom of the form submit button.
 - Header margin fixed to negative 70px.
 - Changed login narration.
-- Removed dupliacte import in config.py
+- Removed duplicate import in config.py
 
 ### Fixed